Changed the low and high range values to Double instead of Long
git-svn-id: https://ncisvn.nci.nih.gov/svn/ctodslabviewer/trunk/Code@245 a393f10d-60de-4571-9f1a-620cac27c168

+++ LabLoader/src/gov/nih/nci/cagrid/labviewer/xml/HL7v3CtLabUnMarshaller.java
@@ -622,7 +622,7 @@ public class HL7v3CtLabUnMarshaller implements UnMarshaller {
 			Node xNode = (Node) xpath.evaluate(expression, doc,
 					XPathConstants.NODE);
 			value = xNode.getNodeValue();
-			clinicalResult.setRefRangeLow(new Long(value));
+			clinicalResult.setRefRangeLow(new Double(value));
 
 		} catch (NullPointerException ex) {
 
@@ -634,7 +634,7 @@ public class HL7v3CtLabUnMarshaller implements UnMarshaller {
 			Node xNode = (Node) xpath.evaluate(expression, doc,
 					XPathConstants.NODE);
 			value = xNode.getNodeValue();
-			clinicalResult.setRefRangeHigh(new Long(value));
+			clinicalResult.setRefRangeHigh(new Double(value));
 
 		} catch (NullPointerException ex) {
